agents_service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
google-genai==0.4.0
redis==5.2.0
structlog==24.4.0
